+++ lib/widgets/index.dart
@@ -1,6 +1,5 @@
 export 'actor_widget.dart';
 export 'app_bar_widget.dart';
-export 'director_and_writerrs_info_widget.dart';
 export 'image_widget.dart';
 export 'image_widget.dart';
 export 'search_widget.dart';
@@ -9,3 +8,5 @@ export 'most_popular_movies_widget.dart';
 export 'movie_item_for_t250.dart';
 export 'movie_item_widget.dart';
 export 'similar_item_widget.dart';
+export 'directors_info_widget.dart';
+export 'writers_info_widget.dart';
